refactor: consolidate machine use cases into data managers and fix measurement types

+++ app/src/main/kotlin/com/jssdvv/ara/machines/domain/type/measurement/TranslationUnits.kt
@@ -5,30 +5,30 @@ import com.jssdvv.ara.R
 
 enum class TranslationUnits(
     @param:StringRes val symbolTextId: Int,
-    val metersPerUnit: Double,
+    val metersPerUnit: Float,
     val millisPerUnit: Int
 ) {
     METERS(
         symbolTextId = R.string.measurement_meters_symbol,
-        metersPerUnit = 1.0,
+        metersPerUnit = 1F,
         millisPerUnit = 1000
     ),
 
     DECIMETERS(
         symbolTextId = R.string.measurement_decimeters_symbol,
-        metersPerUnit = 0.1,
+        metersPerUnit = 0.1F,
         millisPerUnit = 100
     ),
 
     CENTIMETERS(
         symbolTextId = R.string.measurement_centimeters_symbol,
-        metersPerUnit = 0.01,
+        metersPerUnit = 0.01F,
         millisPerUnit = 10
     ),
 
     MILLIMETERS(
         symbolTextId = R.string.measurement_millimeters_symbol,
-        metersPerUnit = 0.001,
+        metersPerUnit = 0.001F,
         millisPerUnit = 1
     )
 }
